requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scipy>=1.10
statsmodels>=0.14
tqdm>=4.66
pyyaml>=6.0
jsonschema>=4.22
rich>=13.7
torch>=2.2
